[feature] discriminator option `config.block`
